Testing only the Service
