+++ Example/Tests/Faro/ServiceSpec.swift
@@ -26,6 +26,8 @@ class ServiceSpec: QuickSpec {
 
                     expect(isInSync).to(beTrue())
                 }
+
+
             }
         }
     }
